fix: common select pyodbc test
