--- API_SERVICE/common_service/database/mixin.py
@@ -1,2 +1,0 @@
-class BaseMixin:
-    ...
